+++ .gitignore
@@ -58,3 +58,4 @@ typings/
 .env
 
 dist
+docs
