+++ jsdoc.json
@@ -1,0 +1,14 @@
+{
+  "source": {
+    "include": "./lib",
+    "includePattern": ".js$"
+  },
+  "opts": {
+    "template": "node_modules/docdash",
+    "destination": "docs/",
+    "readme": "README.md"
+  },
+  "plugins": [
+    "plugins/markdown"
+  ]
+}
