--- lib/core/index.js
@@ -1,2 +1,0 @@
-export {mockConfig} from './mock_config'
-export {process} from './process'
