--- lib/core/mock_config.js
@@ -1,3 +1,0 @@
-export function mockConfig(ctx, name, value) {
-  ctx.config[name] = value
-}
