--- lib/core/process.js
@@ -1,3 +1,0 @@
-export function process(ctx) {
-  return ctx.load().then(() => ctx)
-}
