--- lib/helpers/get_helper.js
@@ -1,3 +1,0 @@
-export function getHelper(ctx, name) {
-  return ctx.extend.helper.get(name)
-}
