--- lib/helpers/has_helper.js
@@ -1,5 +1,0 @@
-import {getHelper} from './get_helper'
-
-export function hasHelper(ctx, name) {
-  return typeof getHelper(ctx, name) !== 'undefined'
-}
