--- lib/helpers/index.js
@@ -1,2 +1,0 @@
-export {getHelper} from './get_helper'
-export {hasHelper} from './has_helper'
