--- lib/routing/has_route.js
@@ -1,3 +1,0 @@
-export function hasRoute(ctx, path) {
-  return ctx.route.get(path) != null
-}
