--- lib/routing/index.js
@@ -1,2 +1,0 @@
-export {contentFor} from './content_for'
-export {hasRoute} from './has_route'
